requirements.txt
Flask==0.10.1
Flask-Inputs==0.2.0
Jinja2>=2.8
MarkupSafe==0.23
Werkzeug==0.11.4
itsdangerous==0.24
jsonschema==2.5.1
six==1.10.0
python-jose==1.3.2
pycodestyle
nose2
gevent>=1.2.2
psutil>=5.4.3
prometheus_client>=0.1.1
netifaces>=0.10.6
msgpack-python==0.4.8
Flask-HTTPAuth>=3.2.3